SystemUI: fix notification light pulse for repeating notifications
additional notifications for apps that already have an
active notification will only send a setPulsing(true)
without a false. This will break all of our assumptions

This is either wanted or a bug. For now instead of searching
for hours to prove who is to blame add a way to handle it.
So if we get another setPulsing(true) while we are
already pulsing dont stop the current one.

Change-Id: I3506a68baf3997b07de053866b3075b0dc7c7962
Signed-off-by: mracar07 <umutcan@umutcanacar.me>